Move restarter-related APIs into dedicated package
