+++ .gitignore
@@ -26,3 +26,4 @@ tmp
 dev
 hack/test
 .DS_Store
+.idea/
